requirements.txt
langchain
langchain_core
langchain_groq
python-dotenv
python-telegram-bot==20.7